+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -22979,7 +22979,9 @@ $MGRF_SelectCompleters = {
 $MGRF_SelectMap = @{
     "Select"=@("Add-MGRFLicense",
                "New-MGRFWorkspace",
+               "New-MGRFWorkspaceApiKey",
                "Remove-MGRFWorkspace",
+               "Remove-MGRFWorkspaceApiKey",
                "Get-MGRFWorkspace",
                "Get-MGRFWorkspaceAuthentication",
                "Remove-MGRFLicense",